backend/requirements.txt
flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21
redis==4.6.0
Flask-Caching==2.1.0
celery==5.3.1
paramiko==3.3.1
netmiko==4.2.0
napalm==4.1.0
ncclient==0.6.15
Jinja2==3.1.2
requests==2.31.0
urllib3==2.0.4
python-dotenv==1.0.0
marshmallow==3.20.1
bcrypt==4.0.1
passlib==1.7.4
python-dateutil==2.8.2
Flask-Mail==0.9.1
Werkzeug==2.3.7
PyJWT==2.8.0
cerberus==1.3.5
prometheus-client==0.17.1
gunicorn==21.2.0
flask-restx==1.1.0
apispec==6.3.0
cryptography==41.0.4